+++ service/simspaceweaver/doc.go
@@ -3,21 +3,20 @@
 // Package simspaceweaver provides the client and types for making API
 // requests to AWS SimSpace Weaver.
 //
-// Amazon Web Services SimSpace Weaver (SimSpace Weaver) is a managed service
-// that you can use to build and operate large-scale spatial simulations in
-// the Amazon Web Services Cloud. For example, you can create a digital twin
-// of a city, crowd simulations with millions of people and objects, and massilvely-multiplayer
-// games with hundreds of thousands of connected players. For more information
-// about SimSpace Weaver, see the Amazon Web Services SimSpace Weaver User Guide
-// (https://docs.aws.amazon.com/simspaceweaver/latest/userguide/) .
+// SimSpace Weaver (SimSpace Weaver) is a managed service that you can use to
+// build and operate large-scale spatial simulations in the Amazon Web Services
+// Cloud. For example, you can create a digital twin of a city, crowd simulations
+// with millions of people and objects, and massively multiplayer games with
+// hundreds of thousands of connected players. For more information about SimSpace
+// Weaver, see the SimSpace Weaver User Guide (https://docs.aws.amazon.com/simspaceweaver/latest/userguide/) .
 //
 // This API reference describes the API operations and data types that you can
 // use to communicate directly with SimSpace Weaver.
 //
 // SimSpace Weaver also provides the SimSpace Weaver app SDK, which you use
 // for app development. The SimSpace Weaver app SDK API reference is included
-// in the SimSpace Weaver app SDK documentation, which is part of the SimSpace
-// Weaver app SDK distributable package.
+// in the SimSpace Weaver app SDK documentation. This documentation is part
+// of the SimSpace Weaver app SDK distributable package.
 //
 // See https://docs.aws.amazon.com/goto/WebAPI/simspaceweaver-2022-10-28 for more information on this service.
 //